Change __init__ to accept dict, NOT json formatted str
My resoning for this is simple.  Looking at the bigger picture, when the worker calls "send()", it receives a json formatted string.  It would like to unpack this json formatted string into an object.  However, it must first determine which object is represented by the json formatted string.  The easist way to do this is to json.loads() the json formatted string back into a dictionary, and then inspect the dictionary's "type" attribute.  So, once worker is ready to instantiate a message object, it has already turned the json formatted string into a dictionary.  Hence, the constructor should take a dictionary